+++ __tests__/tests.config.example.js
