+++ package.json
@@ -41,17 +41,20 @@
   "scripts": {
     "docs": "documentation readme lib/*.js --config documentation.yml --section=API",
     "build": "rollup -c",
-    "dev": "rollup -c --watch"
+    "dev": "rollup -c --watch",
+    "test": "jest"
   },
   "dependencies": {
     "lodash.snakecase": "4.1.1"
   },
   "devDependencies": {
     "@babel/cli": "7.1.5",
-    "@babel/core": "7.1.6",
+    "@babel/core": "^7.2.0",
     "@babel/preset-env": "7.1.6",
     "@babel/preset-flow": "7.0.0",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "10.0.1",
+    "babel-jest": "^23.6.0",
     "documentation": "9.0.0-alpha.1",
     "eslint": "5.9.0",
     "eslint-config-prettier": "3.3.0",
@@ -69,6 +72,7 @@
     "lint-staged": "8.1.0",
     "mathjs": "5.2.3",
     "prettier": "1.15.3",
+    "regenerator-runtime": "^0.13.1",
     "rollup": "0.67.3",
     "rollup-plugin-babel": "4.0.3",
     "rollup-plugin-node-resolve": "3.4.0"
